修改角色和楼栋关系

+++ src/main/java/com/ldps/service/IRoleService.java
@@ -6,6 +6,6 @@ import com.ldps.model.Role;
 public interface IRoleService {
 	List<Role> queryRoleWithPageIndex(String roleName, Integer pageNo, Integer pageSize);
 	int queryCountByCondition(String roleName);
-	int saveOrUpdate(Role role);
+	Long saveOrUpdate(Role role);
 	int deleteById(Long id);
 }
